datatypeFix

+++ BPMConverter.h
@@ -12,8 +12,8 @@
 
 class BPMConverter {
 public:
-	static uint8_t bpmToTimeUnits(uint16_t bpm, uint16_t timeUnitsPerSecond);
-	static uint8_t timeUnitsToBPM(uint16_t timeUnits, uint16_t timeUnitsPerSecond);
+	static uint16_t bpmToTimeUnits(uint16_t bpm, uint16_t timeUnitsPerSecond);
+	static uint16_t timeUnitsToBPM(uint16_t timeUnits, uint16_t timeUnitsPerSecond);
 };
 
 #endif /* BPMCONVERTER_H_ */
